add: 画面ごとにページを分割しロビーを作成
