+++ client/src/pages/game/index.module.css
@@ -1,10 +1,27 @@
-.canvasContainer {
-  width: 100vw;
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
   height: 100svh;
-  overflow: hidden;
 }
 
-.canvasContainer > div {
-  width: 100vw;
-  height: 100vh;
+.container > div {
+  width: 88svmin;
+}
+
+.text {
+  margin-bottom: 1rem;
+  text-align: center;
+}
+
+.buttons {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+}
+
+.buttons > button {
+  width: 8svmin;
+  aspect-ratio: 16 / 9;
 }
